requirements.txt
customtkinter>=5.2.0
pystray>=0.19.0
Pillow>=10.0.0
darkdetect>=0.8.0
matplotlib>=3.8.0
